rest/requirements.txt
Django~=4.2.7
djangorestframework~=3.14.0
bcrypt~=4.0.1
requests~=2.31.0
PyJWT~=2.8.0
psycopg2~=2.9.9